chore(gpum): refactor check test to suite
This commit refactors the check test to a suite so we can share setup
code across tests.
